bill_analyzer/backend/api/requirements.txt
fastapi==0.115.4
bcrypt==4.2.0
redis==5.0.8
prefect==3.0.10
pyjwt==2.8.0
python-multipart==0.0.12
